backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.1,<25.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
